+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v1.5.0-pre
 
 - Add runtime durations to admin pages [#357](https://github.com/CultivateLabs/raif/pull/357)
+- Add support for DeepSeek R1 and v3.2 models via AWS Bedrock. [#358](https://github.com/CultivateLabs/raif/pull/358)
 
 ## v1.4.0